james/requirements.txt
streamlit==1.30.0
pandas==2.1.3
plotly==5.18.0
openpyxl==3.1.2
python-dotenv==1.0.0
clickhouse-driver==0.2.9
python-keycloak==3.7.0
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.7 